Rename internal tests to include .test extension
