+++ testcases/extending-3.test.js
@@ -1,4 +1,4 @@
-import baseTest from './basic.js'
+import baseTest from './basic.test.js'
 import { asyncAssert, repeatTwice, randomSleep } from './extending/reusable-steps.js'
 
 export default {
