+++ testcases/extending-5.test.js
