+++ testcases/repeat-4.test.js
