+++ testcases/repeat-6.test.js
