+++ testcases/repeat-8.test.js
