+++ testcases/repeat-9.test.js
