+++ testcases/sequence-helper.test.js
